refactor websocketdemo to scotlandyard
