+++ app/src/androidTest/java/at/aau/serg/scotlandyard/ExampleInstrumentedTest.kt
@@ -1,10 +1,8 @@
-package at.aau.serg.websocketbrokerdemo
+package at.aau.serg.scotlandyard
 
 import androidx.test.platform.app.InstrumentationRegistry
-import androidx.test.ext.junit.runners.AndroidJUnit4
 
 import org.junit.Test
-import org.junit.runner.RunWith
 
 import org.junit.Assert.*
