+++ app/src/main/java/at/aau/serg/scotlandyard/Callbacks.kt
@@ -1,4 +1,4 @@
-package at.aau.serg.websocketbrokerdemo
+package at.aau.serg.scotlandyard
 
 interface Callbacks {
     fun onResponse(res: String);
