+++ app/src/main/java/at/aau/serg/scotlandyard/MainActivity.kt
@@ -1,4 +1,4 @@
-package at.aau.serg.websocketbrokerdemo
+package at.aau.serg.scotlandyard
 
 import MyStomp
 import android.os.Bundle
