+++ app/src/main/java/at/aau/serg/scotlandyard/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package at.aau.serg.websocketbrokerdemo.ui.theme
+package at.aau.serg.scotlandyard.ui.theme
 
 import androidx.compose.ui.graphics.Color
