+++ app/src/test/java/at/aau/serg/scotlandyard/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package at.aau.serg.websocketbrokerdemo
+package at.aau.serg.scotlandyard
 
 import org.junit.jupiter.api.Test
